+++ pydcs_extensions/irondome/irondome.py
@@ -4,39 +4,9 @@ from game.modsupport import vehiclemod
 
 
 @vehiclemod
-class I9K51_GRAD(VehicleType):
-    id = "I9K51_GRAD"
-    name = "(IDF Mods Project) BM-21 Grad 122mm"
-    detection_range = 0
-    threat_range = 19000
-    air_weapon_dist = 19000
-    eplrs = True
-
-
-@vehiclemod
-class I9K57_URAGAN(VehicleType):
-    id = "I9K57_URAGAN"
-    name = "(IDF Mods Project) Urgan BM-27 220mm"
-    detection_range = 0
-    threat_range = 35800
-    air_weapon_dist = 35800
-    eplrs = True
-
-
-@vehiclemod
-class I9K58_SMERCH(VehicleType):
-    id = "I9K58_SMERCH"
-    name = "(IDF Mods Project) 9A52 Smerch CM 300mm"
-    detection_range = 0
-    threat_range = 70000
-    air_weapon_dist = 70000
-    eplrs = True
-
-
-@vehiclemod
-class IRON_DOME_CP(VehicleType):
-    id = "IRON_DOME_CP"
-    name = "Iron Dome CP"
+class Iron_Dome_David_Sling_CP(VehicleType):
+    id = "Iron_Dome_David_Sling_CP"
+    name = "[IDF Mods] Iron Dome-David Sling CP"
     detection_range = 0
     threat_range = 0
     air_weapon_dist = 0
@@ -46,16 +16,43 @@ class IRON_DOME_CP(VehicleType):
 @vehiclemod
 class IRON_DOME_LN(VehicleType):
     id = "IRON_DOME_LN"
-    name = "Iron Dome LN"
+    name = "[IDF Mods] Iron Dome"
     detection_range = 0
-    threat_range = 25000
-    air_weapon_dist = 25000
+    threat_range = 20000
+    air_weapon_dist = 20000
 
 
 @vehiclemod
-class ELM2048_MMR(VehicleType):
-    id = "ELM2048_MMR"
-    name = "Iron Dome ELM-2048 MMR"
-    detection_range = 412000
+class DAVID_SLING_LN(VehicleType):
+    id = "DAVID_SLING_LN"
+    name = "[IDF Mods] David Sling"
+    detection_range = 0
+    threat_range = 250000
+    air_weapon_dist = 250000
+
+
+@vehiclemod
+class ELM2084_MMR_AD_RT(VehicleType):
+    id = "ELM2084_MMR_AD_RT"
+    name = "[IDF Mods] ELM-2084MMR AD Rotating Mode"
+    detection_range = 475000
+    threat_range = 0
+    air_weapon_dist = 0
+
+
+@vehiclemod
+class ELM2084_MMR_AD_SC(VehicleType):
+    id = "ELM2084_MMR_AD_SC"
+    name = "[IDF Mods] ELM-2084MMR AD Sector Mode"
+    detection_range = 650000
+    threat_range = 0
+    air_weapon_dist = 0
+
+
+@vehiclemod
+class ELM2084_MMR_WLR(VehicleType):
+    id = "ELM2084_MMR_WLR"
+    name = "[IDF Mods] ELM-2084MMR WLR Mode"
+    detection_range = 160000
     threat_range = 0
     air_weapon_dist = 0